Write simple logic for computer choice, player input, and declaring a winner
Simple js functions for getting player input and a random number to represent the computer.
Faced with having to write many conditional statements comparing states in order to determine a winner,
I wanted a simpler solution to the problem. By representing each game choice as a number (0, 1, and 2)
we can substract the computer's state from the user's state to figure out the game outcome. 1 beats
0, 2 beats 1, and 0 beats 2. If the difference is more than 0 (meaning both players chose the same number)
then the sign of the difference (+/-) is used to determine who won. For example, user chooses 0, computer chooses 1.
user - computer == 0 - 1 = -1. Because the difference is not a zero, the outcome is not a draw and because it
is negative, the player lost.